Started written SCB API polling functions: - Changed name of scb.py to scb_wrapper for clarity - Added get(), count() and category() to scb_wrapper - Implemented fetching and filtering of different lists of interest - Added filtered_sni.json to assets (contains all the level 1 codes we are interested in)
