+++ assets/filtered_sni.json
@@ -1,0 +1,178 @@
+[
+    {
+        "VardeLista": [
+            {
+                "Varde": "00",
+                "Text": "Huvudgrupp saknas"
+            },
+            {
+                "Varde": "01",
+                "Text": "Jordbruk och jakt samt service i anslutning härtill"
+            },
+            {
+                "Varde": "02",
+                "Text": "Skogsbruk"
+            },
+            {
+                "Varde": "03",
+                "Text": "Fiske och vattenbruk"
+            },
+            {
+                "Varde": "05",
+                "Text": "Kolutvinning"
+            },
+            {
+                "Varde": "06",
+                "Text": "Utvinning av råpetroleum och naturgas"
+            },
+            {
+                "Varde": "07",
+                "Text": "Utvinning av metallmalmer"
+            },
+            {
+                "Varde": "08",
+                "Text": "Annan utvinning av mineral"
+            },
+            {
+                "Varde": "10",
+                "Text": "Livsmedelsframställning"
+            },
+            {
+                "Varde": "11",
+                "Text": "Framställning av drycker"
+            },
+            {
+                "Varde": "12",
+                "Text": "Tobaksvarutillverkning"
+            },
+            {
+                "Varde": "13",
+                "Text": "Textilvarutillverkning"
+            },
+            {
+                "Varde": "14",
+                "Text": "Tillverkning av kläder"
+            },
+            {
+                "Varde": "15",
+                "Text": "Tillverkning av läder, läder- och skinnvaror m.m."
+            },
+            {
+                "Varde": "16",
+                "Text": "Tillverkning av trä och varor av trä, kork, rotting o.d. utom möbler"
+            },
+            {
+                "Varde": "17",
+                "Text": "Pappers- och pappersvarutillverkning"
+            },
+            {
+                "Varde": "19",
+                "Text": "Tillverkning av stenkolsprodukter och raffinerade petroleumprodukter"
+            },
+            {
+                "Varde": "20",
+                "Text": "Tillverkning av kemikalier och kemiska produkter"
+            },
+            {
+                "Varde": "21",
+                "Text": "Tillverkning av farmaceutiska basprodukter och läkemedel"
+            },
+            {
+                "Varde": "22",
+                "Text": "Tillverkning av gummi- och plastvaror"
+            },
+            {
+                "Varde": "23",
+                "Text": "Tillverkning av andra icke-metalliska mineraliska produkter"
+            },
+            {
+                "Varde": "24",
+                "Text": "Stål- och metallframställning"
+            },
+            {
+                "Varde": "25",
+                "Text": "Tillverkning av metallvaror utom maskiner och apparater"
+            },
+            {
+                "Varde": "26",
+                "Text": "Tillverkning av datorer, elektronikvaror och optik"
+            },
+            {
+                "Varde": "27",
+                "Text": "Tillverkning av elapparatur"
+            },
+            {
+                "Varde": "28",
+                "Text": "Tillverkning av övriga maskiner"
+            },
+            {
+                "Varde": "29",
+                "Text": "Tillverkning av motorfordon, släpfordon och påhängsvagnar"
+            },
+            {
+                "Varde": "30",
+                "Text": "Tillverkning av andra transportmedel"
+            },
+            {
+                "Varde": "31",
+                "Text": "Tillverkning av möbler"
+            },
+            {
+                "Varde": "32",
+                "Text": "Annan tillverkning"
+            },
+            {
+                "Varde": "33",
+                "Text": "Reparation och installation av maskiner och apparater"
+            },
+            {
+                "Varde": "35",
+                "Text": "Försörjning av el, gas, värme och kyla"
+            },
+            {
+                "Varde": "36",
+                "Text": "Vattenförsörjning"
+            },
+            {
+                "Varde": "37",
+                "Text": "Avloppsrening"
+            },
+            {
+                "Varde": "38",
+                "Text": "Avfallshantering; återvinning"
+            },
+            {
+                "Varde": "39",
+                "Text": "Sanering, efterbehandling av jord och vatten samt annan verksamhet för föroreningsbekämpning"
+            },
+            {
+                "Varde": "41",
+                "Text": "Byggande av hus"
+            },
+            {
+                "Varde": "42",
+                "Text": "Anläggningsarbeten"
+            },
+            {
+                "Varde": "43",
+                "Text": "Specialiserad bygg- och anläggningsverksamhet"
+            },
+            {
+                "Varde": "45",
+                "Text": "Handel samt reparation av motorfordon och motorcyklar"
+            },
+            {
+                "Varde": "46",
+                "Text": "Parti- och provisionshandel utom med motorfordon"
+            },
+            {
+                "Varde": "72",
+                "Text": "Vetenskaplig forskning och utveckling"
+            },
+            {
+                "Varde": "95",
+                "Text": "Reparation av datorer, hushållsartiklar och personliga artiklar"
+            }
+        ]
+    }
+]
